// @hey_api/documentsmaterials.swagger/types.gen.ts
export type v2Category = {
  /**
   * Unique identifier of the category.
   */
  categoryId?: string;
  /**
   * Titles of the category in different languages.
   */
  titles?: v2Titles;
  /**
   * Labels assigned to the category.
   */
  labels?: Array<string>;
  /**
   * Multilingual description of the category.
   */
  description?: v2DescriptionJSON;
  /**
   * ID of the user who created the category.
   */
  userId?: string;
  /**
   * number of documents linked to this specific category.
   */
  documentCount?: number;
  /**
   * Timestamp of category creation.
   */
  createdAt?: string;
  /**
   * Timestamp of last update.
   */
  updatedAt?: string;
};
// ...
export type v2DescriptionJSON = {
  en?: string;
  fr?: string;
};
// ...
export type v2DocumentType = {
  /**
   * Unique identifier of the document type.
   */
  id?: string;
  /**
   * List of years with document counts.
   */
  years?: Array<v2Year>;
  /**
   * URL of the image representing the document type.
   */
  imageUrl?: string;
  /**
   * Titles of the document type in different languages.
   */
  titles?: v2Titles;
  /**
   * Caption for the document type.
   */
  caption?: string;
  /**
   * ID of the user who created the document type.
   */
  userId?: string;
  /**
   * Count of documents under this document type.
   */
  documentCount?: string;
  /**
   * Timestamp of document type creation.
   */
  createdAt?: string;
  /**
   * Timestamp of last update.
   */
  updatedAt?: string;
};
// ...
export type v2OwnerNote = {
  /**
   * Message from the owner.
   */
  msg?: string;
  /**
   * Name of the owner.
   */
  name?: string;
  /**
   * Timestamp of the note.
   */
  timestamp?: string;
};

export type v2PublishedDocument = {
  documentId?: string;
  ref?: string;
  documentNumber?: string;
  documentType?: v2DocumentType;
  materialCount?: string;
  documentSuffixNumber?: string;
  language?: string;
  jsonHeader?: string;
  jsonFooter?: string;
  title?: string;
  issueDate?: string;
  updatedAt?: string;
  createdAt?: string;
  userId?: string;
  status?: string;
  documentTypeId?: string;
  rootMaterialId?: string;
  /**
   * Categories this document was published with.
   */
  categories?: Array<v2Category>;
  ownersNotes?: Array<v2OwnerNote>;
};
// ...
export type v2Titles = {
  en?: string;
  fr?: string;
};

/**
 * Response after updating a category.
 */
export type v2UpdateCategoryResponse = {
  category?: v2Category;
};

export type v2Year = {
  /**
   * Year associated with documents.
   */
  year?: number;
  /**
   * Count of documents for the year.
   */
  documentCount?: number;
};